+++ pcs/lib/cib/resource/common.py
@@ -2,9 +2,11 @@ from collections import namedtuple
 from typing import (
     Iterable,
     List,
+    Mapping,
     Optional,
     Set,
     Tuple,
+    cast,
 )
 
 from lxml.etree import _Element
@@ -14,6 +16,7 @@ from pcs.common.reports.item import (
     ReportItem,
     ReportItemList,
 )
+from pcs.common.types import StringCollection
 from pcs.lib.cib import nvpair
 from pcs.lib.cib.resource.bundle import TAG as TAG_BUNDLE
 from pcs.lib.cib.resource.bundle import (
@@ -37,7 +40,10 @@ from pcs.lib.cib.resource.group import (
 from pcs.lib.cib.resource.group import is_group
 from pcs.lib.cib.resource.primitive import TAG as TAG_PRIMITIVE
 from pcs.lib.cib.resource.primitive import is_primitive
-from pcs.lib.cib.tools import ElementSearcher
+from pcs.lib.cib.tools import (
+    ElementSearcher,
+    IdProvider,
+)
 from pcs.lib.xml_tools import find_parent
 
 ALL_RESOURCE_XML_TAGS = sorted(
@@ -45,16 +51,18 @@ ALL_RESOURCE_XML_TAGS = sorted(
 )
 
 
-def are_meta_disabled(meta_attributes):
+def are_meta_disabled(meta_attributes: Mapping[str, str]) -> bool:
     return meta_attributes.get("target-role", "Started").lower() == "stopped"
 
 
-def _can_be_evaluated_as_positive_num(value):
+def _can_be_evaluated_as_positive_num(value: str) -> bool:
     string_wo_leading_zeros = str(value).lstrip("0")
-    return string_wo_leading_zeros and string_wo_leading_zeros[0].isdigit()
+    return (
+        bool(string_wo_leading_zeros) and string_wo_leading_zeros[0].isdigit()
+    )
 
 
-def is_clone_deactivated_by_meta(meta_attributes):
+def is_clone_deactivated_by_meta(meta_attributes: Mapping[str, str]) -> bool:
     return are_meta_disabled(meta_attributes) or any(
         not _can_be_evaluated_as_positive_num(meta_attributes.get(key, "1"))
         for key in ["clone-max", "clone-node-max"]
@@ -64,7 +72,7 @@ def is_clone_deactivated_by_meta(meta_attributes):
 def find_one_resource(
     context_element: _Element,
     resource_id: str,
-    resource_tags: Optional[Iterable[str]] = None,
+    resource_tags: Optional[StringCollection] = None,
 ) -> Tuple[Optional[_Element], ReportItemList]:
     """
     Find a single resource or return None if not found
@@ -85,10 +93,10 @@ def find_one_resource(
 def find_resources(
     context_element: _Element,
     resource_ids: Iterable[str],
-    resource_tags: Optional[Iterable[str]] = None,
+    resource_tags: Optional[StringCollection] = None,
 ) -> Tuple[List[_Element], ReportItemList]:
     """
-    Find a list of resource
+    Find a list of resources
 
     context_element -- an element to be searched in
     resource_id -- id of an element to find
@@ -110,7 +118,8 @@ def find_resources(
 def find_primitives(resource_el: _Element) -> List[_Element]:
     """
     Get list of primitives contained in a given resource
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     if is_bundle(resource_el):
         in_bundle = get_bundle_inner_resource(resource_el)
@@ -198,10 +207,11 @@ def get_parent_resource(resource_el: _Element) -> Optional[_Element]:
     return None
 
 
-def find_resources_to_enable(resource_el):
+def find_resources_to_enable(resource_el: _Element) -> List[_Element]:
     """
     Get resources to enable in order to enable specified resource successfully
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     if is_bundle(resource_el):
         to_enable = [resource_el]
@@ -215,43 +225,42 @@ def find_resources_to_enable(resource_el):
 
     to_enable = [resource_el]
     parent = resource_el.getparent()
-    if is_any_clone(parent) or is_bundle(parent):
+    if parent is not None and (is_any_clone(parent) or is_bundle(parent)):
         to_enable.append(parent)
     return to_enable
 
 
-def enable(resource_el, id_provider):
+def enable(resource_el: _Element, id_provider: IdProvider) -> None:
     """
     Enable specified resource
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     nvpair.arrange_first_meta_attributes(
         resource_el,
-        {
-            "target-role": "",
-        },
+        {"target-role": ""},
         id_provider,
     )
 
 
-def disable(resource_el, id_provider):
+def disable(resource_el: _Element, id_provider: IdProvider) -> None:
     """
     Disable specified resource
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     nvpair.arrange_first_meta_attributes(
         resource_el,
-        {
-            "target-role": "Stopped",
-        },
+        {"target-role": "Stopped"},
         id_provider,
     )
 
 
-def find_resources_to_manage(resource_el):
+def find_resources_to_manage(resource_el: _Element) -> List[_Element]:
     """
-    Get resources to manage to manage the specified resource successfully
-    etree resource_el -- resource element
+    Get resources to set to managed for the specified resource to become managed
+
+    resource_el -- resource element
     """
     # If the resource_el is a primitive in a group, we set both the group and
     # the primitive to managed mode. Otherwise the resource_el, all its
@@ -261,37 +270,41 @@ def find_resources_to_manage(resource_el):
     # as a managed primitive in an unmanaged clone / group is still unmanaged
     # and vice versa.
     res_id = resource_el.attrib["id"]
-    return (
-        [resource_el]  # the resource itself
-        +
-        # its parents
-        find_parent(resource_el, "resources").xpath(
-            # a master or a clone which contains a group, a primitive, or a
-            # grouped primitive with the specified id
-            # OR
-            # a group (in a clone, master, etc. - hence //) which contains a
-            # primitive with the specified id
-            # OR
-            # a bundle which contains a primitive with the specified id
-            """
+    parent_el = []
+    top_element = find_parent(resource_el, {"resources"})
+    if top_element is not None:
+        parent_el = cast(
+            List[_Element],
+            top_element.xpath(
+                # a master or a clone which contains a group, a primitive, or a
+                # grouped primitive with the specified id
+                # OR
+                # a group (in a clone, master, etc. - hence //) which contains a
+                # primitive with the specified id
+                # OR
+                # a bundle which contains a primitive with the specified id
+                """
                 (./master|./clone)[(group|group/primitive|primitive)[@id=$r]]
                 |
                 //group[primitive[@id=$r]]
                 |
                 ./bundle[primitive[@id=$r]]
-            """,
-            r=res_id,
+                """,
+                r=res_id,
+            ),
         )
-        +
-        # its children
-        resource_el.xpath("(./group|./primitive|./group/primitive)")
+    children_el = cast(
+        List[_Element],
+        resource_el.xpath("(./group|./primitive|./group/primitive)"),
     )
+    return [resource_el] + parent_el + children_el
 
 
-def find_resources_to_unmanage(resource_el):
+def find_resources_to_unmanage(resource_el: _Element) -> List[_Element]:
     """
     Get resources to unmanage to unmanage the specified resource successfully
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     # resource hierarchy - specified resource - what to return
     # a primitive - the primitive - the primitive
@@ -343,30 +356,28 @@ def find_resources_to_unmanage(resource_el):
     return []
 
 
-def manage(resource_el, id_provider):
+def manage(resource_el: _Element, id_provider: IdProvider) -> None:
     """
     Set the resource to be managed by the cluster
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     nvpair.arrange_first_meta_attributes(
         resource_el,
-        {
-            "is-managed": "",
-        },
+        {"is-managed": ""},
         id_provider,
     )
 
 
-def unmanage(resource_el, id_provider):
+def unmanage(resource_el: _Element, id_provider: IdProvider) -> None:
     """
     Set the resource not to be managed by the cluster
-    etree resource_el -- resource element
+
+    resource_el -- resource element
     """
     nvpair.arrange_first_meta_attributes(
         resource_el,
-        {
-            "is-managed": "false",
-        },
+        {"is-managed": "false"},
         id_provider,
     )
 
@@ -407,12 +418,14 @@ def find_resources_to_delete(resource_el: _Element) -> List[_Element]:
     return result
 
 
-def validate_move(resource_element, master):
+def validate_move(
+    resource_element: _Element, master: bool
+) -> reports.ReportItemList:
     """
     Validate moving a resource to a node
 
-    etree resource_element -- the resource to be moved
-    bool master -- limit moving to the master role
+    resource_element -- the resource to be moved
+    master -- limit moving to the master role
     """
     report_list = []
     analysis = _validate_move_ban_clear_analyzer(resource_element)
@@ -421,7 +434,7 @@ def validate_move(resource_element, master):
         report_list.append(
             ReportItem.error(
                 reports.messages.CannotMoveResourceBundle(
-                    resource_element.get("id")
+                    str(resource_element.get("id", ""))
                 )
             )
         )
@@ -433,7 +446,7 @@ def validate_move(resource_element, master):
         report_list.append(
             ReportItem.error(
                 reports.messages.CannotMoveResourceClone(
-                    resource_element.get("id")
+                    str(resource_element.get("id", ""))
                 )
             )
         )
@@ -451,7 +464,7 @@ def validate_move(resource_element, master):
         report_list.append(
             ReportItem.error(
                 reports.messages.CannotMoveResourcePromotableInner(
-                    resource_element.get("id"),
+                    str(resource_element.get("id", "")),
                     analysis.promotable_clone_id,
                 )
             )
@@ -460,7 +473,7 @@ def validate_move(resource_element, master):
         report_list.append(
             ReportItem.error(
                 reports.messages.CannotMoveResourceMasterResourceNotPromotable(
-                    resource_element.get("id"),
+                    str(resource_element.get("id", "")),
                     promotable_id=analysis.promotable_clone_id,
                 )
             )
@@ -469,12 +482,14 @@ def validate_move(resource_element, master):
     return report_list
 
 
-def validate_ban(resource_element, master):
+def validate_ban(
+    resource_element: _Element, master: bool
+) -> reports.ReportItemList:
     """
     Validate banning a resource on a node
 
-    etree resource_element -- the resource to be banned
-    bool master -- limit banning to the master role
+    resource_element -- the resource to be banned
+    master -- limit banning to the master role
     """
     report_list = []
     analysis = _validate_move_ban_clear_analyzer(resource_element)
@@ -483,7 +498,7 @@ def validate_ban(resource_element, master):
         report_list.append(
             ReportItem.error(
                 reports.messages.CannotBanResourceMasterResourceNotPromotable(
-                    resource_element.get("id"),
+                    str(resource_element.get("id", "")),
                     promotable_id=analysis.promotable_clone_id,
                 )
             )
@@ -492,22 +507,23 @@ def validate_ban(resource_element, master):
     return report_list
 
 
-def validate_unmove_unban(resource_element, master):
+def validate_unmove_unban(
+    resource_element: _Element, master: bool
+) -> reports.ReportItemList:
     """
     Validate unmoving/unbanning a resource to/on nodes
 
-    etree resource_element -- the resource to be unmoved/unbanned
-    bool master -- limit unmoving/unbanning to the master role
+    resource_element -- the resource to be unmoved/unbanned
+    master -- limit unmoving/unbanning to the master role
     """
     report_list = []
     analysis = _validate_move_ban_clear_analyzer(resource_element)
 
     if master and not analysis.is_promotable_clone:
-        # pylint: disable=line-too-long
         report_list.append(
             ReportItem.error(
                 reports.messages.CannotUnmoveUnbanResourceMasterResourceNotPromotable(
-                    resource_element.get("id"),
+                    str(resource_element.get("id", "")),
                     promotable_id=analysis.promotable_clone_id,
                 )
             )
@@ -532,7 +548,9 @@ class _MoveBanClearAnalysis(
     pass
 
 
-def _validate_move_ban_clear_analyzer(resource_element):
+def _validate_move_ban_clear_analyzer(
+    resource_element: _Element,
+) -> _MoveBanClearAnalysis:
     resource_is_bundle = False
     resource_is_clone = False
     resource_is_in_clone = False
@@ -550,7 +568,9 @@ def _validate_move_ban_clear_analyzer(resource_element):
     elif get_parent_any_clone(resource_element) is not None:
         parent_clone = get_parent_any_clone(resource_element)
         resource_is_in_clone = True
-        if is_master(parent_clone) or is_promotable_clone(parent_clone):
+        if parent_clone is not None and (
+            is_master(parent_clone) or is_promotable_clone(parent_clone)
+        ):
             resource_is_in_promotable_clone = True
             promotable_clone_element = parent_clone
     return _MoveBanClearAnalysis(